pyproject.toml
[tool.poetry]
name = "cloud-sigevent"
version = "2.0.0"
description = "Cloud Sigevent"
authors = [
  "skperez <stepheny.k.perez@jpl.nasa.gov>",
  "jgarde <jgarde@jpl.nasa.gov>"
]
readme = "README.md"
packages = [{include = "podaac/sigevent"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.41"
pydantic = "^2.6.1"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.1"
flake8 = "^7.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
moto = "^4.2.6"
pytest-cov = "^4.1.0"
pylint = "^3.0.3"
mock = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
